requirements.txt
contourpy==1.0.6
cycler==0.11.0
fonttools==4.38.0
keyboard==0.13.5
kiwisolver==1.4.4
matplotlib==3.6.2
numpy==1.24.1
opencv-python==4.6.0.66
packaging==22.0
Pillow==9.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
shiboken6==6.4.1
six==1.16.0
tqdm==4.64.1
yapf==0.32.0
